Modify defintiions for Periodization
