+++ Periodization/CycleExtension.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace YggdrAshill.Ragnarok.Proceduralization
+namespace YggdrAshill.Ragnarok.Periodization
 {
     public static class CycleExtension
     {
