--- Proceduralization/Proceduralization.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <AssemblyName>YggdrAshill.Ragnarok.Proceduralization</AssemblyName>
-    <RootNamespace>YggdrAshill.Ragnarok.Proceduralization</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Periodization\Periodization.csproj" />
-  </ItemGroup>
-
-</Project>
