+++ YggdrAshill.Ragnarok.Specification/Periodization/CycleExtensionSpecification.cs
@@ -1,6 +1,6 @@
 using NUnit.Framework;
 using System;
-using YggdrAshill.Ragnarok.Proceduralization;
+using YggdrAshill.Ragnarok.Periodization;
 
 namespace YggdrAshill.Ragnarok.Specification
 {
