+++ YggdrAshill.Ragnarok.Specification/Periodization/ExecutionSpecification.cs
