+++ YggdrAshill.Ragnarok.Specification/Periodization/OriginationExtensionSpecification.cs
@@ -34,7 +34,7 @@ namespace YggdrAshill.Ragnarok.Specification
         }
 
         [Test]
-        public void CannotBeConvertedWithNull()
+        public void CannotBeConvertedToSpanWithNull()
         {
             Assert.Throws<ArgumentNullException>(() =>
             {
