+++ YggdrAshill.Ragnarok.sln
@@ -16,8 +16,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "YggdrAshill.Ragnarok.Sample
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Periodization", "Periodization\Periodization.csproj", "{E1F4D2E5-E3DE-44B8-A689-712C6BD96D1C}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Proceduralization", "Proceduralization\Proceduralization.csproj", "{368DC7F3-313C-4209-9611-96CC7C630F60}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -40,10 +38,6 @@ Global
 		{E1F4D2E5-E3DE-44B8-A689-712C6BD96D1C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E1F4D2E5-E3DE-44B8-A689-712C6BD96D1C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E1F4D2E5-E3DE-44B8-A689-712C6BD96D1C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{368DC7F3-313C-4209-9611-96CC7C630F60}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{368DC7F3-313C-4209-9611-96CC7C630F60}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{368DC7F3-313C-4209-9611-96CC7C630F60}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{368DC7F3-313C-4209-9611-96CC7C630F60}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
