+++ YggdrAshill.Ragnarok/Periodization/PeriodizationExtension.cs
@@ -32,5 +32,23 @@ namespace YggdrAshill.Ragnarok
 
             return termination.From(new Origination(origination));
         }
+
+        public static ICycle Between(this IExecution execution, Action origination, Action termination)
+        {
+            if (execution is null)
+            {
+                throw new ArgumentNullException(nameof(execution));
+            }
+            if (origination is null)
+            {
+                throw new ArgumentNullException(nameof(origination));
+            }
+            if (termination is null)
+            {
+                throw new ArgumentNullException(nameof(termination));
+            }
+
+            return execution.Between(Origination.Of(origination), Termination.Of(termination));
+        }
     }
 }
