+++ YggdrAshill.Ragnarok/YggdrAshill.Ragnarok.csproj
@@ -6,7 +6,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Periodization\Periodization.csproj" />
-    <ProjectReference Include="..\Proceduralization\Proceduralization.csproj" />
   </ItemGroup>
 
 </Project>
